refactor: rename package for #775 tests
