+++ tests/issue-775-duplicate-service-port/src/test/java/io/dekorate/example/Issue775TwoPortsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.dekorate.annotationless;
+package io.dekorate.example;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -23,17 +23,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.test.context.TestPropertySource;
 
 import io.dekorate.utils.Serialization;
-import io.fabric8.kubernetes.api.model.HasMetadata;
-import io.fabric8.kubernetes.api.model.KubernetesList;
-import io.fabric8.kubernetes.api.model.Service;
-import io.fabric8.kubernetes.api.model.ServicePort;
+import io.fabric8.kubernetes.api.model.*;
 
-public class Issue775OnePortTest {
+@TestPropertySource(locations = "./application2.properties")
+public class Issue775TwoPortsTest {
 
   @Test
-  public void shouldHaveOnePort() {
+  public void shouldHaveUniquePortNumberForTwoPorts() {
     KubernetesList list = Serialization
         .unmarshalAsList(getClass().getClassLoader().getResourceAsStream("META-INF/dekorate/kubernetes.yml"));
     assertNotNull(list);
@@ -41,12 +40,17 @@ public class Issue775OnePortTest {
     assertNotNull(s);
 
     List<ServicePort> p = s.getSpec().getPorts();
-    assertEquals(1, p.size());
+    assertEquals(2, p.size());
 
     ServicePort servicePort1 = p.get(0);
     assertEquals("http", servicePort1.getName());
     assertEquals(80, servicePort1.getPort());
     assertEquals(8080, servicePort1.getTargetPort().getIntVal());
+
+    ServicePort servicePort2 = p.get(1);
+    assertEquals("https", servicePort2.getName());
+    assertEquals(443, servicePort2.getPort());
+    assertEquals(8443, servicePort2.getTargetPort().getIntVal());
   }
 
   <T extends HasMetadata> Optional<T> findFirst(KubernetesList list, Class<T> t) {
