+++ tests/issue-775-sb-server-port/src/main/java/io/dekorate/example/DemoApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.dekorate.annotationless;
+package io.dekorate.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
