+++ tests/issue-775-sb-server-port/src/test/java/io/dekorate/example/Issue775SpringBootServerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.dekorate.annotationless;
+package io.dekorate.example;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -23,16 +23,17 @@ import java.util.List;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.test.context.TestPropertySource;
 
 import io.dekorate.utils.Serialization;
-import io.fabric8.kubernetes.api.model.*;
+import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.fabric8.kubernetes.api.model.KubernetesList;
+import io.fabric8.kubernetes.api.model.Service;
+import io.fabric8.kubernetes.api.model.ServicePort;
 
-@TestPropertySource(locations = "./application2.properties")
-public class Issue775TwoPortsTest {
+public class Issue775SpringBootServerTest {
 
   @Test
-  public void shouldHaveUniquePortNumberForTwoPorts() {
+  public void shouldHaveSpringBootServerPort() {
     KubernetesList list = Serialization
         .unmarshalAsList(getClass().getClassLoader().getResourceAsStream("META-INF/dekorate/kubernetes.yml"));
     assertNotNull(list);
@@ -40,17 +41,12 @@ public class Issue775TwoPortsTest {
     assertNotNull(s);
 
     List<ServicePort> p = s.getSpec().getPorts();
-    assertEquals(2, p.size());
+    assertEquals(1, p.size());
 
     ServicePort servicePort1 = p.get(0);
     assertEquals("http", servicePort1.getName());
     assertEquals(80, servicePort1.getPort());
-    assertEquals(8080, servicePort1.getTargetPort().getIntVal());
-
-    ServicePort servicePort2 = p.get(1);
-    assertEquals("https", servicePort2.getName());
-    assertEquals(443, servicePort2.getPort());
-    assertEquals(8443, servicePort2.getTargetPort().getIntVal());
+    assertEquals(8090, servicePort1.getTargetPort().getIntVal());
   }
 
   <T extends HasMetadata> Optional<T> findFirst(KubernetesList list, Class<T> t) {
